Bring framework back to life
Removed some Obj-C deprecated warnings.
Replace some deprecated Redland functions with their newer counterparts.
Update linking so the test suite runs again; have to see how this works for ARM code later on.

+++ Classes/RedlandNode.m
@@ -137,7 +137,7 @@
 - (id)initWithCoder:(NSCoder *)coder
 {
     unsigned char const *buffer;
-    unsigned int bufSize;
+    NSUInteger bufSize;
     librdf_node *node;
     NSParameterAssert(coder != nil);
     
@@ -217,9 +217,9 @@
         return NO;
 }
 
-- (unsigned)hash
+- (NSUInteger)hash
 {
-    return (unsigned)wrappedObject;
+    return (NSUInteger)wrappedObject;
 }
 
 - (NSComparisonResult)compare:(id)otherNode
